Manage roles in views

+++ backoffice_operateurs/views/ads.py
@@ -2,20 +2,24 @@
 from backoffice_operateurs import db
 from backoffice_operateurs.forms.taxis import ADSCreateForm, ADSUpdateForm
 from backoffice_operateurs.models import taxis as taxis_models
-from backoffice_operateurs.views.zupc import zupc_autocomplete
 from flask import Blueprint, render_template, request, redirect, url_for, abort
 from backoffice_operateurs.utils import create_obj_from_json
 from flask import render_template, request, redirect, url_for, abort, jsonify
-from flask.ext.security import login_required
+from flask.ext.security import login_required, current_user
 from datetime import datetime
 
 
 mod = Blueprint('ads', __name__)
 
 def ads_list():
+    if not taxis_models.ADS.can_be_listed_by(current_user):
+        abort(403)
+    q = taxis_models.ADS.query
+    if not current_user.has_role('admin'):
+        q.filter_by(added_by=current_user.id)
     page = int(request.args.get('page')) if 'page' in request.args else 1
     return render_template('lists/ads.html',
-        ads_list=taxis_models.ADS.query.paginate(page))
+        ads_list=q.paginate(page))
 
 def ads_create_api():
     json = request.get_json()
@@ -33,13 +37,16 @@ def ads_create_api():
     return jsonify(new_ads.as_dict())
 
 def ads_details(immatriculation, numero):
-    ads = taxis_models.ADS.query.filter_by(numero=numero,
-            immatriculation=immatriculation).all()
+    filters = {
+            "numero": numero,
+            "immatriculation": immatriculation
+            }
+    ads = taxis_models.ADS.query.filter_by(**filters).all()
     if not ads:
         abort(404)
     ads = ads[0]
     d = taxis_models.ADS.__dict__
-    keys_to_show = ["numero", "marque", "modele", "immatriculation"]
+    keys_to_show = ads.showable_fields(current_user)
     is_valid_key = lambda k: hasattr(k, "info") and k.info.has_key("label") and k.info['label'] and k.key in keys_to_show
     #@TODO: make it dependent of the user's role
     if request.content_type == "application/json":
@@ -72,6 +79,8 @@ def ads_form():
         ads = taxis_models.ADS.query.get(request.args.get("id"))
         if not ads:
             abort(404)
+        if not ads.can_be_edited_by(current_user)
+            abort(403)
         form = ADSUpdateForm(obj=ads, zupc=ads.ZUPC.nom)
     else:
         form = ADSCreateForm()
@@ -81,13 +90,13 @@ def ads_form():
             form.populate_obj(ads)
             db.session.add(ads)
             db.session.commit()
-            return redirect(url_for('ads'))
+            return redirect(url_for('ads.ads'))
         elif ads:
             ads.last_update_at = datetime.now().isoformat()
             form.populate_obj(ads)
             if form.validate():
                 db.session.commit()
-                return redirect(url_for('ads'))
+                return redirect(url_for('ads.ads'))
     return render_template('forms/ads.html', form=form)
 
 @mod.route('/ads/delete')
@@ -98,6 +107,8 @@ def ads_delete():
     ads = taxis_models.ADS.query.get(request.args.get("id"))
     if not ads:
         abort(404)
+    if not ads.can_be_deleted_by(current_user):
+        abort(403)
     db.session.delete(ads)
     db.session.commit()
-    return redirect(url_for("ads"))
+    return redirect(url_for("ads.ads"))
