+++ backoffice_operateurs/views/conducteur.py
@@ -6,7 +6,7 @@ from backoffice_operateurs.models import taxis as taxis_models
 from flask import Blueprint, render_template, request, redirect, url_for, abort
 from flask import render_template, request, redirect, url_for, abort, jsonify,\
         current_app
-from flask.ext.security import login_required
+from flask.ext.security import login_required, current_user
 from backoffice_operateurs.utils import create_obj_from_json
 
 mod = Blueprint('conducteur', __name__)
@@ -30,6 +30,8 @@ def conducteur_api_add():
 
 
 def conducteurs_list():
+    if not taxis_models.Conducteur.can_be_listed_by(current_user):
+        abort(403)
     page = int(request.args.get('page')) if 'page' in request.args else 1
     return render_template('lists/conducteurs.html',
         conducteur_list=taxis_models.Conducteur.query.paginate(page))
@@ -53,6 +55,8 @@ def conducteur_form():
         conducteur = taxis_models.Conducteur.query.get(request.args.get("id"))
         if not conducteur:
             abort(404)
+        if not conducteur.can_be_edited_by(current_user):
+            abort(403)
         form = ConducteurUpdateForm(obj=conducteur)
     else:
         form = ConducteurCreateForm()
@@ -62,13 +66,13 @@ def conducteur_form():
             form.populate_obj(conducteur)
             if form.validate():
                 db.session.commit()
-                return redirect(url_for('conducteurs'))
+                return redirect(url_for('conducteurs.conducteurs'))
         else:
             conducteur = taxis_models.Conducteur()
             form.populate_obj(conducteur)
             db.session.add(conducteur)
             db.session.commit()
-            return redirect(url_for('conducteurs'))
+            return redirect(url_for('conducteurs.conducteurs'))
     return render_template('forms/conducteur.html', form=form,
         form_method="POST", submit_value="Modifier")
 
@@ -81,6 +85,8 @@ def conducteur_delete():
     conducteur = taxis_models.Conducteur.query.get(request.args.get("id"))
     if not conducteur:
         abort(404)
+    if not conducteur.can_be_deleted_by(current_user):
+        abort(403)
     db.session.delete(conducteur)
     db.session.commit()
-    return redirect(url_for("conducteurs"))
+    return redirect(url_for("conducteurs.conducteurs"))
